requirements.txt
flet==0.22.*
llama-cpp-python